% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cn_to_ssm.R
\name{assign_cn_to_ssm}
\alias{assign_cn_to_ssm}
\title{Assign CN to SSM.}
\usage{
assign_cn_to_ssm(
  this_sample_id,
  genes,
  this_seq_type = "genome",
  projection = "grch37",
  coding_only = FALSE,
  assume_diploid = FALSE,
  include_silent = FALSE,
  ...
)
}
\arguments{
\item{this_sample_id}{Sample ID of the sample you want to annotate.}

\item{genes}{Genes of interest.}

\item{this_seq_type}{Specified seq type for returned data.}

\item{projection}{specified genome projection that returned data is in reference to.}

\item{coding_only}{Optional. set to TRUE to restrict to only coding variants.}

\item{assume_diploid}{Optional. If no local seg file is provided, instead of defaulting to a GAMBL sample, this parameter annotates every mutation as copy neutral.}

\item{include_silent}{Logical parameter indicating whether to include silent mutations into coding mutations. Default is FALSE}
}
\value{
A list containing a data frame (MAF-like format) with two extra columns:
log.ratio is the log ratio from the seg file (NA when no overlap was found)
as well as the segmented copy number data with the same copy number information
CN is the rounded absolute copy number estimate of the region based on log.ratio (NA when no overlap was found)
}
\description{
Annotate mutations with their copy number information.
}
\details{
This function takes a sample ID with the \code{this_sample_id} parameter and annotates mutations with copy number information.
A variety of parameters are at hand for a customized workflow. For example, the user can specify if only coding mutations are of interest.
To do so, set \code{coding_only = TRUE}. This function internally calls \code{get_ssm_by_samples} and \code{get_sample_cn_segments}.
This function can also take a vector with genes of interest (\code{genes}) that the returned data frame will be restricted to.
}
\examples{
cn_list = assign_cn_to_ssm(this_sample_id = "DOHH-2",
                           coding_only = TRUE)

}
