% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding_ssm.R
\name{bind_genomic_data}
\alias{bind_genomic_data}
\title{Bind maf data together}
\usage{
bind_genomic_data(...)
}
\arguments{
\item{...}{All maf_data or seg_data objects to be combined}
}
\value{
data.frame
}
\description{
Combine multiple maf_data objects and retain metadata such as genome_build. This function
will not allow you to combine maf_data objects that have different genome_build values. An error will also
be thrown if the same sample id is found in more than one of the inputs
}
\examples{

merged_maf = bind_genomic_data(maf1,maf2)


}
