% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_excess_params.R
\name{check_excess_params}
\alias{check_excess_params}
\title{Check Excess Params}
\usage{
check_excess_params(...)
}
\arguments{
\item{...}{Parameters to check.}
}
\value{
Nothing
}
\description{
Function for checking excessive parameter names.
This function will notify the user if any unavailable parameters are called for any given given function.
This function is designed to work as internal function-call in already available GAMBLR functions.
}
\details{
Catch function calls containing unsupported arguments.
}
