% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coding_class}
\alias{coding_class}
\title{Coding Class}
\format{
\subsection{\code{coding_class}}{

A vector of characters with 12 elements.
\describe{
\item{coding_class}{Different coding classes.}
}
}
}
\usage{
coding_class
}
\description{
A vector with characters detailing different coding classes.
}
\keyword{datasets}
