% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour_codes}
\alias{colour_codes}
\title{Colour Codes}
\format{
\subsection{\code{colour_codes}}{

A data frame with 221 rows and 5 columns.
\describe{
\item{category}{Describes category for any given colour.}
\item{group}{Describes the group for any given colour.}
\item{name}{The name for any given colour.}
\item{colour}{Colour annotated in HEX format.}
\item{is_alias}{Describes if the colour has an alias (yes) or not (NA)}
}
}
}
\usage{
colour_codes
}
\description{
A data frame with colour codes (hex) arranged into different categories, groups.
}
\keyword{datasets}
