% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding_ssm.R
\name{create_maf_data}
\alias{create_maf_data}
\title{Create MAF Data}
\usage{
create_maf_data(maf_df, genome_build)
}
\arguments{
\item{maf_df}{A data frame containing the MAF data.}

\item{genome_build}{A string specifying the genome build ("grch37" or "hg38").}
}
\value{
A data frame with class attributes for MAF data.
}
\description{
This function creates MAF (Mutation Annotation Format) data from the given input.
}
