% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cn_segments.R
\name{create_seg_data}
\alias{create_seg_data}
\title{Create Segmented Data}
\usage{
create_seg_data(seg_df, genome_build)
}
\arguments{
\item{seg_df}{A data frame containing the segmented data.}

\item{genome_build}{A string specifying the genome build ("grch37" or "hg38").}
}
\value{
A data frame with class attributes for segmented data.
}
\description{
This function creates segmented data from the given input.
}
\examples{
seg_df <- data.frame(...)
create_seg_data(seg_df, "grch37")
}
