% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytobands_grch37}
\alias{cytobands_grch37}
\title{Cytobands coordinates (grch37)}
\format{
\subsection{\code{cytobands_grch37}}{

A data frame with 862 rows and 5 columns
\describe{
\item{cb.chromosome}{Chromosome of the cytoband.}
\item{cb.start}{Start position of the cytoband.}
\item{cb.end}{End position of the cytoband.}
\item{cb.name}{Cytoband name.}
\item{label}{Cytoband label.}
}
}
}
\usage{
cytobands_grch37
}
\description{
A data frame in bed format with coordinates of cytobands relative to grch37.
}
\keyword{internal}
