% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytobands_hg38}
\alias{cytobands_hg38}
\title{Cytobands coordinates (hg38)}
\format{
\subsection{\code{cytobands_hg38}}{

A data frame with 862 rows and 5 columns
\describe{
\item{cb.chromosome}{Chromosome of the cytoband.}
\item{cb.start}{Start position of the cytoband.}
\item{cb.end}{End position of the cytoband.}
\item{cb.name}{Cytoband name.}
\item{label}{Cytoband label.}
}
}
}
\usage{
cytobands_hg38
}
\description{
A data frame in bed format with coordinates of cytobands relative to hg38.
}
\keyword{internal}
