% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlbclass_vector}
\alias{dlbclass_vector}
\title{DLBclass Features Vector.}
\format{
\subsection{\code{dlbclass_vector}}{

A character vector of length 163. Names are features, values are their DLBclass subtype association. Values are a factor with levels consistent with typical subtype ordering. Useful for splitGeneGroups argument to prettyOncoplot.

N.B. Uses DLBclass feature nomenclature, e.g. "MYD88.L265P". May not map automatically to Hugo_Symbol.
}
}
\usage{
dlbclass_vector
}
\description{
A named factor of features used in DLBclass with their subtype association.
}
\keyword{internal}
