% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMBLR.data-package.R
\docType{package}
\name{GAMBLR.data-package}
\alias{GAMBLR.data}
\alias{GAMBLR.data-package}
\title{GAMBLR.data: Collection of Curated Data for Genomic Analysis of Mature B-cell Lymphomas in R}
\description{
The package contains manually curated data for the genomic Analysis of mature B-cell lymphomas in R, such as regions of somatic hypermutation, lymphoma genes, etc.
}
\author{
\strong{Maintainer}: Kostiantyn Dreval \email{kdreval@sfu.ca} (\href{https://orcid.org/0000-0002-6214-2843}{ORCID})

Authors:
\itemize{
  \item Ryan Morin \email{rdmorin@sfu.ca} (\href{https://orcid.org/0000-0003-2932-7800}{ORCID})
  \item Adam Mattsson \email{cmattsson@bcgsc.ca} (\href{https://orcid.org/0000-0002-6318-7912}{ORCID})
}

}
\keyword{internal}
