% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode_to_symbol}
\alias{gencode_to_symbol}
\title{ENSG to symbol converter}
\format{
\subsection{\code{gencode_to_symbol}}{

A data frame with 60038 rows and 3 columns
\describe{
\item{ensembl_gene_id}{ENSEMBL gene id in a format without .<number>.}
\item{gene_id}{ENSEMBL gene id in a format containing .<number>.}
\item{hgnc_symbol}{Human-readable gene symbol.}
}
}
}
\usage{
gencode_to_symbol
}
\description{
A data frame with ENSEMBL gene ids in different format and the corresponding
hgnc symbols.
}
\keyword{internal}
