% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ashm_count_matrix.R
\name{get_ashm_count_matrix}
\alias{get_ashm_count_matrix}
\title{Get ASHM Count Matrix.}
\usage{
get_ashm_count_matrix(
  regions_bed,
  these_samples_metadata,
  this_seq_type,
  projection = "grch37"
)
}
\arguments{
\item{regions_bed}{A bed file with one row for each region.}

\item{these_samples_metadata}{This is used to complete your matrix. All GAMBL
samples will be used by default. Provide a data frame with at least
sample_id for all samples if you are using non-GAMBL data.}

\item{this_seq_type}{The seq type to return results for. Only used if no
metadata is provided with these_samples_metadata.}

\item{projection}{Which genome build to use for the mutations
(must match the coordinate system your regions to avoid a nonsense result)}
}
\value{
matrix
}
\description{
Prepare a matrix with one row per sample and one column per
region using a set of hypermutated regions.
}
\details{
Values are the number of mutations in that patient in the region.
}
\examples{
regions_bed = create_bed_data(GAMBLR.data::grch37_ashm_regions,
                              fix_names="concat",
                              concat_cols=c("gene","region"),
                              sep="-")
my_meta = get_gambl_metadata() \%>\% dplyr::filter(pathology=="DLBCL")
matrix <- get_ashm_count_matrix(
     regions_bed = regions_bed,
     this_seq_type = "genome"
)

#this example should fail because the regions_bed is not hg38
 matrix <- get_ashm_count_matrix(regions_bed=regions_bed,
                            this_seq_type = "genome",
                            these_samples_metadata = my_meta,
                            projection = "hg38")
# Error in get_ashm_count_matrix(
# Your projection argument does not match the genome_build of regions_bed

# format the name column to include the coordinates instead of the gene
regions_bed = create_bed_data(GAMBLR.data::hg38_ashm_regions,
                           fix_names="concat",
                           concat_cols=c("chr_name","hg38_start","hg38_end"),
                           sep="-")

 matrix_hg38 <- get_ashm_count_matrix(regions_bed=regions_bed,
                                      this_seq_type = "genome",
                                      these_samples_metadata = my_meta,
                                      projection = "hg38")
print(dim(matrix_hg38))
print(head(matrix_hg38[,c(1:8)]))
}
