% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding_ssm.R
\name{get_coding_ssm}
\alias{get_coding_ssm}
\title{Get Coding SSMs}
\usage{
get_coding_ssm(
  limit_cohort,
  exclude_cohort,
  limit_pathology,
  limit_samples,
  these_samples_metadata,
  force_unmatched_samples,
  projection = "grch37",
  seq_type,
  basic_columns = TRUE,
  maf_cols = NULL,
  from_flatfile = NULL,
  augmented = NULL,
  min_read_support = 3,
  groups = NULL,
  include_silent = TRUE,
  engine = NULL
)
}
\arguments{
\item{limit_cohort}{Supply this to restrict mutations to one or more cohorts in a vector.}

\item{exclude_cohort}{Supply this to exclude mutations from one or more cohorts in a vector.}

\item{limit_pathology}{Supply this to restrict mutations to one pathology.}

\item{limit_samples}{Supply this to restrict mutations to a vector of sample_id (instead of subsetting using the provided metadata)}

\item{these_samples_metadata}{Supply a metadata table to auto-subset the data to samples in that table before returning}

\item{force_unmatched_samples}{Optional argument for forcing unmatched samples, using \link{get_ssm_by_samples}.}

\item{projection}{Reference genome build for the coordinates in the MAF file. The default is hg19 genome build.}

\item{seq_type}{The seq_type you want back, default is genome.}

\item{basic_columns}{Set to FALSE to override the default behavior of returning only the first 45 columns of MAF data.}

\item{maf_cols}{if basic_columns is set to FALSE, the user can specify what columns to be returned within the MAF. This parameter can either be a vector of indexes (integer) or a vector of characters (matching columns in MAF).}

\item{from_flatfile}{This parameter does not do anything for this version of get_manta_sv. See \link[GAMBLR.results:get_coding_ssm]{GAMBLR.results::get_coding_ssm} for more info.}

\item{augmented}{This parameter does not do anything for this version of get_manta_sv. See \link[GAMBLR.results:get_coding_ssm]{GAMBLR.results::get_coding_ssm} for more info.}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs)}

\item{groups}{This parameter does not do anything for this version of get_manta_sv. See \link[GAMBLR.results:get_coding_ssm]{GAMBLR.results::get_coding_ssm} for more info.}

\item{include_silent}{Logical parameter indicating whether to include silent mutations into coding mutations. Default is TRUE.}

\item{engine}{This parameter does not do anything for this version of get_manta_sv. See \link[GAMBLR.results:get_coding_ssm]{GAMBLR.results::get_coding_ssm} for more info.}
}
\value{
A data frame (.maf) with coding SSMs in reference to the selected projection.
}
\description{
Convenience function for loading coding Simple Somatic Mutations (SSM) from the bundled data \link{sample_data}.
}
\details{
This "bare bones" function was developed to retrieve coding SSM calls for non-GSC-users.
For users with GSC access it is highly recommended to instead call \link[GAMBLR.results:get_coding_ssm]{GAMBLR.results::get_coding_ssm}.
Effectively retrieve coding SSM calls. Multiple filtering parameters are available for this function.
For more information on how to implement the filtering parameters, refer to the parameter descriptions as well as examples in the vignettes.
It relies on the bundled sample data in this package.
}
\examples{
#return SSMs in reference to GRCh37:
ssm_grch37 = get_coding_ssm()

#return SSMs in reference to hg38:
ssm_hg38 = get_coding_ssm(this_projection = "hg38")

}
