% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding_ssm_status.R
\name{get_coding_ssm_status}
\alias{get_coding_ssm_status}
\title{Get Coding SSM Status.}
\usage{
get_coding_ssm_status(
  gene_symbols,
  these_samples_metadata,
  maf_data,
  include_hotspots = TRUE,
  keep_multihit_hotspot = FALSE,
  review_hotspots = TRUE,
  genes_of_interest = c("FOXO1", "MYD88", "CREBBP"),
  genome_build = "hg19",
  include_silent = FALSE,
  include_silent_genes,
  ...
)
}
\arguments{
\item{gene_symbols}{A vector of gene symbols for which the mutation status
will be tabulated. If not provided, lymphoma genes will be returned
by default.}

\item{these_samples_metadata}{The metadata for samples of interest to be
included in the returned matrix. Only the column "sample_id" is
required. If not provided, the example metadata is used as default.}

\item{maf_data}{data frame in maf format. Must be in the grch37 projection.}

\item{include_hotspots}{Logical parameter indicating whether hotspots object
should also be tabulated. Default is TRUE.}

\item{keep_multihit_hotspot}{Logical parameter indicating whether to keep the
gene annotation as mutated when the gene has both hot spot and
non-hotspot mutation. Default is FALSE. If set to TRUE, will report the
number of non-hotspot mutations instead of tabulating for just mutation
presence.}

\item{review_hotspots}{Logical parameter indicating whether hotspots object
should be reviewed to include functionally relevant mutations or rare
lymphoma-related genes. Default is TRUE.}

\item{genes_of_interest}{A vector of genes for hotspot review. Currently only
FOXO1, MYD88, and CREBBP are supported.}

\item{genome_build}{Reference genome build for the coordinates in the MAF
file. The default is hg19 genome build.}

\item{include_silent}{Logical parameter indicating whether to include silent
mutations into coding mutations. Default is FALSE.}
}
\value{
A data frame with tabulated mutation status.
}
\description{
Tabulate mutation status (SSM) for a set of genes.
}
\details{
This function takes a vector of gene symbols and subsets the
incoming MAF to specified genes. If no genes are provided, the function will
default to all lymphoma genes.  Currently only the grch37 genome build is
supported for hotspot annotation and review for this version of the function.
}
\examples{
coding_tabulated_df = get_coding_ssm_status(
 maf_data = GAMBLR.data::sample_data$grch37$maf,
 gene_symbols = "EGFR"
)

#all lymphoma genes from bundled NHL gene list
coding_tabulated_df = get_coding_ssm_status()

}
