% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{get_genome_build}
\alias{get_genome_build}
\title{Get Genome Build}
\usage{
get_genome_build(data)
}
\arguments{
\item{data}{A data frame with genome build attribute.}
}
\value{
A string specifying the genome build.
}
\description{
This function retrieves the genome build attribute from the data.
}
