% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_manta_sv.R
\name{get_manta_sv}
\alias{get_manta_sv}
\title{Get Manta SVs}
\usage{
get_manta_sv(
  these_sample_ids,
  these_samples_metadata,
  projection = "grch37",
  chromosome,
  qstart,
  qend,
  region,
  pairing_status,
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{these_sample_ids}{A vector of multiple sample_id (or a single sample ID as a string) that you want results for.}

\item{these_samples_metadata}{A metadata table (with sample IDs in a column) to auto-subset the data to samples in that table before returning.}

\item{projection}{The projection genome build. Default is grch37.}

\item{chromosome}{Optional, the chromosome you are restricting to (can be prefixed or not prefixed).}

\item{qstart}{Optional, query start coordinate of the range you are restricting to.}

\item{qend}{Optional, query end coordinate of the range you are restricting to.}

\item{region}{Optional, region formatted like chrX:1234-5678 (chromosome can be prefixed or not prefixed) instead of specifying chromosome, start and end separately.}

\item{pairing_status}{Use to restrict results (if desired) to matched or unmatched results (default is to return all). This parameter takes the filtering condition as a string ("matched" or "unmatched").}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned. Default is 0.1.}

\item{min_score}{The lowest Manta somatic score for a SV to be returned. Default is 40.}

\item{pass}{If TRUE (default) only return SVs that are annotated with PASS in the FILTER column. Set to FALSE to keep all variants, regardless if they PASS the filters.}

\item{verbose}{Set to FALSE to minimize the output to console. Default is TRUE. This parameter also dictates the verbosity of any helper function internally called inside the main function.}

\item{...}{Any additional parameters.}
}
\description{
Convenience function for retrieving Manta Structural Variants (SVs) from the bundled data \link{sample_data}.
}
\details{
This "bare bones" function was developed to retrieve Manta SV calls bundled within GAMBLR.data.
To obtain SV calls for multiple samples, give \code{these_sample_ids} a vector of sample IDs.
Alternatively, the user can also provide the \code{these_samples_metadata} parameter to make use of an already subset metadata table.
In this case, the returned SVs will be restricted to the sample_ids within that data frame.
This function internally calls \link{id_ease} to streamline sample ID/metadata parameters.
This function can also restrict the returned calls to any genomic regions specified within \code{chromosome}, \code{qstart}, \code{qend},
or the complete region specified under \code{region} (in chr:start-end format), note that chromosome can be either prefixed or not prefixed.
Useful filtering parameters are also available, use \code{min_vaf} to set the minimum tumour VAF for a SV to be returned and \code{min_score}
to set the lowest Manta somatic score for a SV to be returned. \code{pair_status} can be used to return variants from either matched or unmatched samples.
In addition, the user can chose to return all variants, even the ones not passing the filter criteria. To do so, set \code{pass = FALSE} (default is TRUE).
}
\examples{
#load packages
library(dplyr)

#lazily get every SV in the table with default quality filters
all_sv = get_manta_sv()

#get all SVs DLBCL cell line samples
cell_line_meta = GAMBLR.data::sample_data$meta \%>\% 
  dplyr::filter(cohort == "DLBCL_cell_lines")
  
dlbcl_sv = get_manta_sv(these_samples_metadata = cell_line_meta)

#get the SVs in a region around MYC
myc_locus_sv = get_manta_sv(region = "8:128723128-128774067")

}
