% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssm_by_patients.R
\name{get_ssm_by_patients}
\alias{get_ssm_by_patients}
\title{Get SSM By Patients.}
\usage{
get_ssm_by_patients(
  these_patient_ids,
  these_samples_metadata,
  projection = "grch37",
  this_seq_type = "genome",
  min_read_support = 3,
  basic_columns = TRUE,
  maf_cols = NULL,
  ...
)
}
\arguments{
\item{these_patient_ids}{A vector of patient IDs that you want results for.
The user can also use a metadata table that has been subset to the patient IDs of interest (\code{these_samples_metadata}).}

\item{these_samples_metadata}{A metadata subset to contain the rows corresponding to the patients of interest.
If the vector of patient IDs is missing (\code{these_patient_ids}), this function will default to all patient IDs in the metadata table given to this parameter.}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38). Default is grch37.}

\item{this_seq_type}{The seq type you want results for. Default is "genome".}

\item{min_read_support}{Subset returned variants to a set minimum read support. Default si 3.}

\item{basic_columns}{Return first 45 columns of MAF rather than full details. Default is TRUE.}

\item{maf_cols}{if basic_columns is set to FALSE, the user can specify what columns to be returned within the MAF.
This parameter can either be a vector of indexes (integer) or a vector of characters (matching columns in MAF).}

\item{...}{Any additional parameters.}
}
\value{
A data frame with SSM calls for the selected patients in MAF format.
}
\description{
Get MAF-format data frame for more than one patient.
}
\details{
This function returns variants from a set of patients.
This function internally calls \link{get_ssm_by_samples}.
Thus, the main contents of this function is to wrangel the provided patient IDs,
so that the corresponding sample IDs can be provided to the internal call of \code{get_sssm_by_samples}
This function expects either a vector of patient IDs (\code{these_patients_ids})
or an already subset metadata table (\code{these_samples_metadata}).
}
\examples{
#load packages
library(dplyr)

#basic usage, these_patient_ids
my_patient = get_ssm_by_patients(these_patient_ids = "DOHH-2")

#using a subset metadata tablee to retreive patient SSMs
cell_line_meta = GAMBLR.data::sample_data$meta \%>\% 
 dplyr::filter(cohort == "DLBCL_cell_lines")
 
patient_maf = get_ssm_by_patients(these_samples_metadata = cell_line_meta, 
                                  this_seq_type = "genome")

}
