% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssm_by_region.R
\name{get_ssm_by_region}
\alias{get_ssm_by_region}
\title{Get SSM By Region.}
\usage{
get_ssm_by_region(
  chromosome,
  qstart,
  qend,
  region = "",
  streamlined = FALSE,
  projection = "grch37",
  seq_type = "genome",
  min_read_support = 3,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{chromosome}{The chromosome you are restricting to (with or without a chr prefix).}

\item{qstart}{Query start coordinate of the range you are restricting to.}

\item{qend}{Query end coordinate of the range you are restricting to.}

\item{region}{Region formatted like chrX:1234-5678 instead of specifying chromosome, start and end separately.}

\item{streamlined}{Return Start_Position and Tumor_Smaple_Barcode as the only two MAF columns. Default is FALSE.}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38).}

\item{seq_type}{The seq_type you want back, default is genome. Currently only genome is supported.}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs).}

\item{verbose}{Set to FALSE to prevent ANY message to be printed.
In most cases, this parameter should be left to TRUE.
The parameter was added to accommodate for noisy output
when running this function in a loop for retrieving SSM
for multiple regions \link{get_ssm_by_regions}.}

\item{...}{Any additional parameters.}
}
\value{
A data frame containing all mutations (MAF) in the specified region.
}
\description{
Retrieve all SSMs from the GAMBL database within a single genomic coordinate range.
}
\details{
This function lets the user specify a region of interest for returning SSM calls within that region.
There are multiple ways a region can be specified. For example, the user can provide the full region in a "region" format (chr:start-end) to the \code{region} parameter.
Or, the user can provide chromosome, start and end coordinates individually with \code{chr}, \code{start}, and \code{end} parameters.
}
\examples{
my_mutations = get_ssm_by_region(region = "chr8:128,723,128-128,774,067")

#specifying chromosome, start and end individually
my_mutations = get_ssm_by_region(chromosome = "8",
                                 qstart = 128723128,
                                 qend = 128774067)

}
