% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch37_ashm_regions}
\alias{grch37_ashm_regions}
\title{GRCh37 ASHM Regions}
\format{
\subsection{\code{grch37_ashm_regions}}{

A data frame with 129 rows and 6 columns.
\describe{
\item{chr_name}{Chromsome name.}
\item{hg19_start}{Start coordinate for region in respect to hg19.}
\item{hg19_end}{End coordinate for region in respect to hg19.}
\item{gene}{Gene symbol in Hugo format.}
\item{region}{Region name.}
\item{regulatory_comment}{Annotates region with regulatory information.}
}
}
}
\usage{
grch37_ashm_regions
}
\description{
Aberrant Somatic Hyper Mutation (ASHM) regions in respect to GRCh37.
}
\keyword{datasets}
