% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_ashm_regions}
\alias{hg38_ashm_regions}
\title{hg38 ASHM Regions}
\format{
\subsection{\code{hg38_ashm_regions}}{

A data frame with 129 rows and 6 columns.
\describe{
\item{chr_name}{Chromsome name.}
\item{hg38_start}{Start coordinate for region in respect to hg38.}
\item{hg38_end}{End coordinate for region in respect to hg38.}
\item{gene}{Gene symbol in Hugo format.}
\item{region}{Region name.}
\item{regulatory_comment}{Annotates region with regulatory information.}
}
}
}
\usage{
hg38_ashm_regions
}
\description{
Aberrant Somatic Hyper Mutation (ASHM) regions in respect to hg38.
}
\keyword{datasets}
