% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hotspots_annotations}
\alias{hotspots_annotations}
\title{Hotspot Annotations}
\format{
\subsection{\code{hotspots_annotations}}{

A data frame with 170 rows and 4 columns
\describe{
\item{MAX_COORD}{Coordinate of hotspot region position identified by HotMaps.}
\item{Start_Position}{Coordinate of gene position.}
\item{Chromosome}{Name of gene chromosome.}
\item{hot_spot}{Hot spot annotation.}
}
}
}
\usage{
hotspots_annotations
}
\description{
A data frame with high-quality positions of ssm hotspots in selected genes.
This resource is based on GAMBL data and was used for hotspot annotation in
cFL/dFL classifier.
}
\keyword{internal}
