% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphgen_genes}
\alias{lymphgen_genes}
\title{LymphGen Genes.}
\format{
\subsection{\code{lymphgen_genes}}{

A character vector of length 86. Names are genes, values are their LymphGen subtype association. Values are a factor with levels consistent with typical subtype ordering. Useful for splitGeneGroups argument to prettyOncoplot.
}
}
\usage{
lymphgen_genes
}
\description{
A named factor of genes used in LymphGen with their subtype association.
}
\keyword{internal}
