% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma_genes_bl_v_latest}
\alias{lymphoma_genes_bl_v_latest}
\title{Lymphoma Genes BL Latest}
\format{
\subsection{\code{lymphoma_genes_bl_v_latest}}{

A data frame withh 111 rows and 9 columns.
\describe{
\item{Gene}{Gene ID in Hugo format.}
\item{Tier}{Tier of gene in this pathology.}
\item{aSHM}{Whether this gene is associated with aSHM.}
\item{QC}{QC of a particular gene.}
\item{Mean.Variant.Quality}{Mean variant quality.}
\item{citekey}{Alphanumeric representation of the citekey where this gene was first described.}
\item{PMID}{Pubmed ID to associated study.}
\item{MutationEffect}{Annotates the effect of the gene mutation.}
\item{Mutation.PMID}{Pubmed ID to associated study where mutation effect is described.}
}
}
}
\usage{
lymphoma_genes_bl_v_latest
}
\description{
Genes frequently associated with Burkitt Lymphoma (BL). This is the most up-to-date version of this dataset.
}
\keyword{datasets}
