% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma_genes_dlbcl_v0.2}
\alias{lymphoma_genes_dlbcl_v0.2}
\title{Lymphoma Genes DLBCL v0.2}
\format{
\subsection{\code{lymphoma_genes_dlbcl_v0.2}}{

A data frame withh 335 rows and 9 columns.
\describe{
\item{Gene}{Gene ID in Hugo format.}
\item{Tier}{Tier of gene in this pathology.}
\item{aSHM}{Whether this gene is associated with aSHM.}
\item{QC}{QC of a particular gene.}
\item{Mean.Variant.Quality}{Mean variant quality.}
\item{citekey}{Alphanumeric representation of the citekey where this gene was first described.}
\item{PMID}{Pubmed ID to associated study.}
\item{MutationEffect}{Annotates the effect of the gene mutation.}
\item{Mutation.PMID}{Pubmed ID to associated study where mutation effect is described.}
\item{MutationEffect.citekey}{Alphanumeric representation of the citekey to associated study where mutation effect is described.}
\item{Mutation.PMID}{Whether this gene is a feature in LymphGen classifier.}
}
}
}
\usage{
lymphoma_genes_dlbcl_v0.2
}
\description{
Genes frequently associated with Diffuse large B cell lymphoma (DLBCL). This is version 0.2.
}
\keyword{internal}
