% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma_genes_lymphoma_genes_v0.0}
\alias{lymphoma_genes_lymphoma_genes_v0.0}
\title{Lymphoma Genes v0.0}
\format{
\subsection{\code{lymphoma_genes_lymphoma_genes_v0.0}}{

A data frame with 196 rows and 12 columns.
\describe{
\item{Gene}{Gene symbol in Hugo format.}
\item{DLBCL}{Boolean varaible annotating if the event (genne) is associated with DLBCL.}
\item{FL}{Boolean varaible annotating if the event (genne) is associated with FL.}
\item{BL}{Boolean varaible annotating if the event (genne) is associated with BL.}
\item{MCL}{Boolean varaible annotating if the event (genne) is associated with MCL.}
\item{CLL}{Boolean varaible annotating if the event (genne) is associated with CLL.}
\item{ensembl_geene_id}{Gene ID in ensembl format.}
\item{hgnc_symbol}{Gene symbol in HGNC format.}
\item{LymphGen}{Boolean variable stating if the event is a described lymphgen or not.}
\item{Reddy}{Boolean variable stating if the event is described in the study (Reddy).}
\item{Chappuy}{Boolean variable stating if the event is described in the study (Chappuy).}
\item{entrezgene_id}{Gene ID in entrez fromat.}
}
}
}
\usage{
lymphoma_genes_lymphoma_genes_v0.0
}
\description{
A comprehenssive resource of genes associated with different types of lymphomas.
}
\keyword{datasets}
