% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma_genes_mcl_v_latest}
\alias{lymphoma_genes_mcl_v_latest}
\title{Lymphoma Genes MCL Latest}
\format{
\subsection{\code{lymphoma_genes_mcl_v_latest}}{

A data frame with 53 rows and 13 columns.
\describe{
\item{ensembl_gene_id}{Gene ID in ensembl format.}
\item{Gene}{Gene symbol in Hugo format.v}
\item{curated}{Boolean variable annotating if the event is currated or not.}
\item{Bea}{Boolean variable stating if the event is described in the study (Bea).}
\item{Zhang}{Boolean variable stating if the event is described in the study (Zhang).}
\item{Pararajalingam}{Boolean variable stating if the event is described in the study (Pararajalingam).}
\item{Nadeu}{Boolean variable stating if the event is described in the study (Nadeu).}
\item{Other_support}{Variable that annotates the event if there are other support available.}
\item{appx_overall_freq}{Approximate overall frequency.}
\item{gambl_freq}{Frequency of event described in GAMBL.}
\item{common_alias}{Variable annotating other common aliases for the event, if such exists.}
\item{noncoding_driver_support}{Boolean variable annotating if the event has noncoding driver support or not.}
\item{aSHM}{Boolean varaible annotating if the event is considered an aSHM or not.}
}
}
}
\usage{
lymphoma_genes_mcl_v_latest
}
\description{
Genes frequently associated with Mantle cell lymphoma (MCL).
#' This is the most up-to-date version of this dataset.
}
\keyword{datasets}
