% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding_ssm.R
\name{preserve_genomic_attributes}
\alias{preserve_genomic_attributes}
\title{Preserve Genomic Attributes}
\usage{
preserve_genomic_attributes(new_data, old_data)
}
\arguments{
\item{new_data}{A data frame resulting from dplyr operations.}

\item{old_data}{The original data frame with genomic attributes.}
}
\value{
A data frame with preserved genomic attributes.
}
\description{
This function preserves the genomic attributes and class after dplyr operations.
}
