% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_domains}
\alias{protein_domains}
\title{Protein Domains}
\format{
\subsection{\code{protein_domains}}{

A data frame with 92849 rows and 12 columns
\describe{
\item{HGNC}{HUGO Gene Nomenclature Committee.}
\item{refseq.ID}{Reference sequence identifier.}
\item{protein.ID}{Protein identifier.}
\item{aa.length}{Amino acid length.}
\item{Start}{Coordinate of amino acid position.}
\item{End}{Coordinate of amino acid position.}
\item{domain.source}{Domain source.}
\item{Label}{Abbreviated domain type.}
\item{domain.anno}{Domain type.}
\item{pfam}{Protein family.}
\item{Description}{Description of the domain.}
\item{Description}{NA.}
}
}
}
\usage{
protein_domains
}
\description{
A data frame with high-quality positions of amino acid positions in their
corresponding domains.
}
\keyword{internal}
