% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_to_chunks.R
\name{region_to_chunks}
\alias{region_to_chunks}
\title{Separate a chromosome region into chunks}
\usage{
region_to_chunks(region)
}
\arguments{
\item{region}{A single string that stores a chromosome region. Any format like
"chr1:100000-200000", "1:100000-200000", "chr1:100'000-200'000" is possible.}
}
\value{
A list with length 3 and names "chromosome", "start" and "end.
}
\description{
\code{region_to_chunks} breaks the input string that stores a chromosome
region to create a list with chromosome number and start and end positions as
separated elements.
}
\examples{
region_to_chunks(region = "chr1:100000-200000")

}
