% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{somatic_hypermutation_locations_GRCh37_v0.6}
\alias{somatic_hypermutation_locations_GRCh37_v0.6}
\title{Somatic Hypermutation Locations GRCh37 v0.6}
\format{
\subsection{\code{somatic_hypermutation_locations_GRCh37_v0.6}}{

A data frame with 127 rows and 6 columns.
\describe{
\item{chr_name}{Chromsome name.}
\item{hg19_start}{Start coordinate for region in respect to hg19.}
\item{hg19_end}{End coordinate for region in respect to hg19.}
\item{gene}{Gene symbol in Hugo format.}
\item{region}{Region name.}
\item{regulatory_comment}{Annotates region with regulatory information.}
}
}
}
\usage{
somatic_hypermutation_locations_GRCh37_v0.6
}
\description{
A data frame with somatic hypermutation locations in respect to GRCh37, version 0.6.
}
\keyword{internal}
