% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{somatic_hypermutation_locations_GRCh38_v0.0}
\alias{somatic_hypermutation_locations_GRCh38_v0.0}
\title{Somatic Hypermutation Locations GRCh38 v0.0}
\format{
\subsection{\code{somatic_hypermutation_locations_GRCh38_v0.0}}{

A data frame with 76 rows and 7 columns.
\describe{
\item{chr_name}{Chromsome name.}
\item{hg38_start}{Start coordinate for region in respect to hg19.}
\item{hg38_end}{End coordinate for region in respect to hg19.}
\item{gene}{Gene symbol in Hugo format.}
\item{region}{Region name.}
\item{regulatory_comment}{Annotates region with regulatory information.}
\item{name}{Location name.}
}
}
}
\usage{
somatic_hypermutation_locations_GRCh38_v0.0
}
\description{
A data frame with somatic hypermutation locations in respect to GRCh38, version 0.0.
}
\keyword{internal}
