% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{strip_genomic_classes}
\alias{strip_genomic_classes}
\title{Strip Genomic Data Classes}
\usage{
strip_genomic_classes(x, classes = c("genomic_data", "maf_data", "bed_data"))
}
\arguments{
\item{x}{An object, such as one of your genomic data objects.}

\item{classes}{A character vector of class names to remove. The default is
c("genomic_data", "maf_data", "bed_data").}
}
\value{
The object with the specified classes removed.
}
\description{
This function removes custom classes associated with genomic data objects
(by default, "genomic_data", "maf_data", and "bed_data") from the class attribute
of an object. This can be useful when you want to revert an S3 object to its
underlying data.frame (or data.table) classes without converting the object.
}
\keyword{internal}
