% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_targetscan.R
\name{target_targetscan}
\alias{target_targetscan}
\title{Integrate miRNA (MicroRNA) data.}
\usage{
target_targetscan(maf, mirna_target, projection = "grch37")
}
\arguments{
\item{maf}{Required. MAF data frame (required columns: Chromosome, Start_Position, End_Position).}

\item{mirna_target}{Data frame (required columns: Chromosome, Start_Position, End_Position, miRNA, sites)}

\item{projection}{The genome build projection for the variants you are working with (default is grch37)}
}
\value{
data frame in MAF format with mirna and sites columns.
}
\description{
Add miRNA data to MAF file.
}
\details{
This function considers the positions of the variants and looks for
mirnas targets that are in the same positions to add their names and sites to
them.
}
\examples{
sample_df = target_targetscan (maf)

}
