% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_hotspots.R
\name{annotate_hotspots}
\alias{annotate_hotspots}
\title{Annotate Hotspots.}
\usage{
annotate_hotspots(mutation_maf, ...)
}
\arguments{
\item{mutation_maf}{A data frame in MAF format.}

\item{...}{Any other parameter. These parameters will be ignored.}
}
\value{
The same data frame with one additional column "hot_spot".
}
\description{
Annotate MAF-like data frome with a hot_spot column indicating recurrent mutations.
}
\details{
This function takes an already loaded MAF data frame with the \code{mutation_maf} parameter.
}
\examples{
my_metadata = get_gambl_metadata()
all_coding_ssm = get_coding_ssm(these_samples_metadata = my_metadata,
                                projection = "grch37",
                                this_seq_type = "genome")

hot_ssms = annotate_hotspots(all_coding_ssm)

}
