% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cn_to_ssm.R
\name{assign_cn_to_ssm}
\alias{assign_cn_to_ssm}
\title{Assign CN to SSM.}
\usage{
assign_cn_to_ssm(
  this_sample_id,
  genes,
  this_seq_type = "genome",
  projection = "grch37",
  coding_only = FALSE,
  assume_diploid = FALSE,
  include_silent = FALSE,
  ...
)
}
\arguments{
\item{this_sample_id}{Sample ID of the sample you want to annotate.}

\item{genes}{A vector of characters with gene symbols (Hugo).}

\item{this_seq_type}{Specified seq type for returned data. Default is genome.}

\item{projection}{Specified genome projection that returned data is in
reference to. Default is grch37.}

\item{coding_only}{Optional. Set to TRUE to restrict to only coding variants
(ssm). Deafult is FALSE.}

\item{assume_diploid}{Optional, this parameter annotates every mutation as
copy neutral. Default is FALSE.}

\item{include_silent}{Logical parameter indicating whether to include silent
mutations into coding mutations. Default is FALSE. This parameter only
makes sense if \code{coding_only} is set to TRUE.}

\item{...}{Any additional parameters.}
}
\value{
A list containing a data frame (MAF-like format) with three extra
columns:
- log.ratio is the log ratio from the seg file (NA when no overlap).
- LOH
- CN (the rounded absolute copy number estimate of the region based on
log.ratio, NA when no overlap was found).
}
\description{
Annotate mutations with their copy number information.
}
\details{
This function takes a sample ID with the \code{this_sample_id} parameter
and annotates mutations with copy number information. A variety of
parameters are at hand for a customized workflow. For example,
the user can specify if only coding mutations are of interest. To do so,
set \code{coding_only = TRUE}. This function internally calls
\code{get_ssm_by_samples} and \code{get_sample_cn_segments}. This function can
also take a vector with genes of interest (\code{genes}) that the returned
data frame will be restricted to.
}
\examples{
cn_list = assign_cn_to_ssm(
     this_sample_id = "DOHH-2",
     coding_only = TRUE
)

}
