% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{bind_genomic_data}
\alias{bind_genomic_data}
\title{Bind maf or other genomic data together}
\usage{
bind_genomic_data(..., check_id = TRUE)
}
\arguments{
\item{...}{All maf_data or seg_data objects to be combined.}

\item{check_id}{Logical. If TRUE (the default), the function will check for the presence of the expected ID column
and for duplicate sample IDs across the inputs. Set to FALSE to skip this check.}
}
\value{
data.frame with combined data and preserved genome_build metadata.
}
\description{
Combine multiple maf_data objects and retain metadata such as genome_build.
This function will not allow you to combine maf_data objects that have different genome_build values.
An error will also be thrown if the same sample id is found in more than one of the inputs (if check_id is TRUE).
}
\examples{

merged_maf = bind_genomic_data(maf1, maf2,check_id=FALSE)

}
\keyword{internal}
