% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlbcl90_genes}
\alias{dlbcl90_genes}
\title{DLBCL90 genes}
\format{
\subsection{\code{dlbcl90_genes}}{

A data frame with 73 rows and 6 columns
\describe{
\item{Gene.symbol}{Human-readable gene symbol.}
\item{coef}{Relative weight of the gene in DLBCL90 classification.}
\item{Assay}{Name of the assay.}
\item{ensembl_gene_id}{ENSEMBL gene id.}
\item{gene_id}{ENSEMBL gene id with version.}
\item{hgnc_symbol}{Human-readable gene symbol.}
}
}
}
\usage{
dlbcl90_genes
}
\description{
A data frame with genes and their weights for DLBCL90.
}
\keyword{datasets}
