% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding_ssm.R
\name{get_coding_ssm}
\alias{get_coding_ssm}
\title{Get Coding SSMs}
\usage{
get_coding_ssm(
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  projection = "grch37",
  this_seq_type = "genome",
  tool_name = "slms-3",
  basic_columns = TRUE,
  min_read_support = 3,
  include_silent = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{these_sample_ids}{Optional, a vector of multiple sample_id (or a single
sample ID as a string) that you want results for.}

\item{these_samples_metadata}{Optional, a metadata table (with sample IDs in
a column) to subset the return to. If not provided (and if
\code{these_sample_ids} is not provided), the function will return all
samples from the specified seq_type in the metadata.}

\item{projection}{Reference genome build for the coordinates in the MAF file.
The default is grch37.}

\item{this_seq_type}{The this_seq_type you want back, default is genome.}

\item{tool_name}{Optionally specify which tool to report variant from. The
default is slms-3, also supports "publication" to return the exact
variants as reported in the original papers.}

\item{basic_columns}{Set to FALSE to override the default behavior of
returning only the first 45 columns of MAF data.}

\item{min_read_support}{Only returns variants with at least this many reads
in t_alt_count.}

\item{include_silent}{Logical parameter indicating whether to include silent
mutations into coding mutations. Default is TRUE.}

\item{verbose}{Set to FALSE to minimize the output to console. Default is
TRUE. This parameter also dictates the verbosity of any helper function
internally called inside the main function.}

\item{...}{Any additional parameters.}
}
\value{
data frame
}
\description{
Convenience function for loading coding Simple Somatic Mutations
(SSM) from the bundled data \link{sample_data}.
}
\details{
This "bare bones" function was developed to retrieve coding SSM
calls for non-GSC-users. Effectively retrieve coding SSM calls. Multiple
filtering parameters are available for this function. For more
information on how to implement the filtering parameters, refer to the
parameter descriptions as well as examples in the vignettes. This
function depends on the bundled sample data in this package.
}
\examples{

 # Get mutations from exome data originally aligned to grch37
ssm_exomes_grch37 = get_coding_ssm(projection = "grch37",this_seq_type = "capture")

# Get mutations from genome data, hg38 build
ssm_genomes_hg38 = get_coding_ssm(projection = "hg38",this_seq_type = "genome")




}
