% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_genes}
\alias{get_genes}
\title{Get genes from one or more gene lists}
\usage{
get_genes(
  entities = c("DLBCL", "MCL", "BL"),
  curated_only = TRUE,
  gene_format = "symbol"
)
}
\arguments{
\item{entities}{Optional vector specifying one or more lymphoma entities
e.g. MCL, DLBCL, BL.}

\item{curated_only}{Specify FALSE to retrieve all genes or leave default
for the curated subset.}

\item{gene_format}{Specify what to return as output. Can be one of:
* "symbol" (the default): list of gene symbols
* "ensembl": list of ENSEMBLE IDs
* "data.frame": data frame with column Gene and per-entity gene status}
}
\value{
A character vector of gene symbol or Ensembl IDs or a data frame.
}
\description{
Retrieve gene names from bundled lymphoma gene lists.
}
\details{
Complete lists of genes described as significantly mutated in large
lymphoma studies have been curated and provided with this package.
}
\examples{
all_dlbcl_genes <- get_genes(entities = "DLBCL", curated_only = FALSE)
all_curated_genes <- get_genes()
}
