% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_cn_segments.R
\name{get_sample_cn_segments}
\alias{get_sample_cn_segments}
\title{Get Sample CN Segments.}
\usage{
get_sample_cn_segments(
  this_sample_id,
  multiple_samples = FALSE,
  sample_list,
  from_flatfile = NULL,
  projection = "grch37",
  this_seq_type = "genome",
  with_chr_prefix = FALSE,
  streamlined = FALSE
)
}
\arguments{
\item{this_sample_id}{Optional argument, single sample_id for the sample to retrieve segments for.}

\item{multiple_samples}{Set to TRUE to return cn segments for multiple samples specified in \code{samples_list} parameter. Default is FALSE.}

\item{sample_list}{Optional vector of type character with one sample per row, required if multiple_samples is set to TRUE.}

\item{from_flatfile}{This parameter does not do anything for this version of get_manta_sv. See \link[GAMBLR.results:get_sample_cn_segments]{GAMBLR.results::get_sample_cn_segments} for more info.}

\item{projection}{Selected genome projection for returned CN segments. Default is "grch37".}

\item{this_seq_type}{Seq type for returned CN segments. One of "genome" (default) or "capture".}

\item{with_chr_prefix}{Set to TRUE to add a chr prefix to chromosome names. Default is FALSE.}

\item{streamlined}{Return a minimal output rather than full details. Default is FALSE.}
}
\value{
A data frame of segments for a specific or multiple sample ID(s).
}
\description{
Get all segments for a single (or multiple) sample_id(s).
}
\details{
This function returns CN segments for samples. This works for single sample or multiple samples.
For multiple samples, remember to set the Boolean parameter \code{multiple_samples = TRUE} and give the \code{sample_lsit} a vector of characters with one sample ID per row.
For more information on how this function can be run in different ways, refer to the parameter descriptions, examples and vignettes.
Is this function not what you are looking for? Try one of the following, similar, functions; \link[GAMBLR:assign_cn_to_ssm]{GAMBLR::assign_cn_to_ssm}, \link[GAMBLR:get_cn_segments]{GAMBLR::get_cn_segments}, \link[GAMBLR:get_cn_states]{GAMBLR::get_cn_states},
}
\examples{
#return sample CNN segmetns for one sample
my_segs = get_sample_cn_segments(this_sample_id = "DOHH-2", projection = "hg38", streamlined = TRUE)

#define some samples
my_samples = c("SU-DHL-4", "DOHH-2")

#return CN segments for thhese samples
these_segs = get_sample_cn_segments(multiple_samples = TRUE,
                                    sample_list = my_samples)

}
