% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssm_by_patients.R
\name{get_ssm_by_patients}
\alias{get_ssm_by_patients}
\title{Get SSM By Patients.}
\usage{
get_ssm_by_patients(
  these_patient_ids,
  these_samples_metadata,
  projection = "grch37",
  this_seq_type = "genome",
  tool_name = "slms-3",
  this_study,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{these_patient_ids}{A vector of patient IDs that you want results for.
The user can also use a metadata table that has been subset to the patient IDs of interest (see \code{these_samples_metadata}).}

\item{these_samples_metadata}{A metadata subset to contain the rows corresponding to the patients of interest.
If the vector of patient IDs is missing (\code{these_patient_ids}), this function will default to all patient IDs in the metadata table given to this parameter.}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38). Default is grch37.}

\item{this_seq_type}{The seq type you want results for. Default is "genome".}

\item{tool_name}{Optionally specify which tool to report variant from. The default is slms-3, also supports "publication" to return the exact variants as reported in the original papers.}

\item{this_study}{Optionally specify first name of the author for the paper
from which the variants should be returned for.
This parameter can either be a vector of indexes (integer) or a vector of characters (matching columns in MAF).}

\item{verbose}{Set to FALSE to minimize the output to console. Default is TRUE. This parameter also dictates the verbosity of any helper function internally called inside the main function.}

\item{...}{Any additional parameters.}
}
\value{
A data frame with SSM calls for the selected patients in MAF format.
}
\description{
Get MAF-format data frame for more than one patient.
}
\details{
This function returns variants from a set of patients.
This function internally calls \link{get_ssm_by_samples}.
Thus, the main contents of this function is to wrangle the provided patient IDs,
so that the corresponding sample IDs can be provided to the internal call of \code{get_ssm_by_samples}.
This function expects either a vector of patient IDs (\code{these_patients_ids})
or an already subset metadata table (\code{these_samples_metadata}).
}
\examples{
#load packages
library(dplyr)

#basic usage, these_patient_ids
my_patient = get_ssm_by_patients(these_patient_ids = "DOHH-2")

#using a subset metadata tablee to retreive patient SSMs
cell_line_meta = GAMBLR.data::sample_data$meta \%>\%
 dplyr::filter(cohort == "DLBCL_cell_lines")

patient_maf = get_ssm_by_patients(these_samples_metadata = cell_line_meta,
                                  this_seq_type = "genome")

}
