% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssm_by_regions.R
\name{get_ssm_by_regions}
\alias{get_ssm_by_regions}
\title{Get SSM By Regions.}
\usage{
get_ssm_by_regions(
  regions_vector,
  regions_bed,
  streamlined = TRUE,
  basic_columns = FALSE,
  use_name_column = FALSE,
  projection = "grch37",
  min_read_support = 3,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{regions_vector}{Either provide a vector of regions in the chr:start-end format.
If this parameter is provided, the function will auto set \code{use_name_column} to FALSE.}

\item{regions_bed}{Provide a data frame in BED format with the coordinates you want to retrieve (recommended).
This parameter can also accept an additional column with region names that will be added to the return if \code{use_name_column = TRUE}}

\item{streamlined}{If set to TRUE (default) only 3 columns will be kept in the returned data frame (start, sample_id and region_name).
Set to FALSE to return 6 columns. See \code{basic_columns} for more information on what columns to include in the return.}

\item{basic_columns}{Set to TRUE for returning a maf with standarad 45 columns. Default is FALSE. If TRUE, \code{streamlined} and \code{use_name_column} will be ignored.}

\item{use_name_column}{If your bed-format data frame has a name column (must be named "name") these can be used to name your regions.
This parameter. is nott accepted if \code{regions_vector} is provided.}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38).}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs).}

\item{verbose}{Set to TRUE to maximize the output to console. Default is TRUE.
This parameter also dictates the verbosity of any helper function internally called inside the main function.}

\item{...}{Any additional parameters.}
}
\value{
Returns a data frame of variants in MAF-like format.
}
\description{
Efficiently retrieve all mutations across a range of genomic regions.
}
\details{
This function internally calls \link{get_ssm_by_region} to retrieve SSM calls for the specified regions.
}
\examples{
#basic usage, adding custom names from bundled ashm data frame
regions_bed = dplyr::mutate(GAMBLR.data::grch37_ashm_regions, 
  name = paste(gene, region, sep = "_"))

ashm_basic_details = get_ssm_by_regions(regions_bed = regions_bed,
                                        streamlined = FALSE,
                                        use_name_column = TRUE)

}
