% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssm_by_regions.R
\name{get_ssm_by_regions}
\alias{get_ssm_by_regions}
\title{Get SSM By Regions.}
\usage{
get_ssm_by_regions(
  these_sample_ids,
  these_samples_metadata,
  this_seq_type = "genome",
  regions_list,
  regions_bed,
  streamlined = TRUE,
  use_name_column = FALSE,
  projection = "grch37",
  min_read_support = 3,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{these_sample_ids}{Optional, a vector of multiple sample_id (or a single sample ID as a string) that you want results for.}

\item{these_samples_metadata}{Optional, a metadata table (with sample IDs in a column) to subset the return to.
If not provided (and if \code{these_sample_ids} is not provided), the function will return all samples from the specified seq_type in the metadata.}

\item{this_seq_type}{The this_seq_type you want back, default is genome.}

\item{regions_list}{A vector of regions in the chr:start-end format to restrict the returned SSM calls to.}

\item{regions_bed}{A data frame in BED format with the coordinates you want to retrieve (recommended).
This parameter can also accept an additional column with region names that will be added to the return if \code{use_name_column = TRUE}}

\item{streamlined}{If set to TRUE (default) only 3 columns will be kept in the returned data frame (start, sample_id and region_name).}

\item{use_name_column}{If your bed-format data frame has a name column (must be named "name") these can be used to name your regions in the returned data frame.}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38), default is grch37.}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs).}

\item{verbose}{Set to TRUE to maximize the output to console. Default is TRUE.
This parameter also dictates the verbosity of any helper function internally called inside the main function.}

\item{...}{Any additional parameters.}
}
\value{
Returns a data frame of variants in MAF-like format.
}
\description{
Efficiently retrieve all mutations across a range of genomic regions.
}
\details{
This function internally calls \link{get_ssm_by_region} to retrieve SSM calls for the specified regions.
}
\examples{
#basic usage, adding custom names from bundled ashm data frame
regions_bed = dplyr::mutate(GAMBLR.data::grch37_ashm_regions, 
  name = paste(gene, region, sep = "_"))

ashm_basic_details = get_ssm_by_regions(regions_bed = regions_bed,
                                        streamlined = FALSE,
                                        use_name_column = TRUE)

}
