% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssm_by_samples.R
\name{get_ssm_by_samples}
\alias{get_ssm_by_samples}
\title{Get SSM By Samples.}
\usage{
get_ssm_by_samples(
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  this_seq_type = "genome",
  projection = "grch37",
  tool_name = "slms-3",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{these_sample_ids}{A vector of one or more sample IDs that you want results for.}

\item{these_samples_metadata}{Optional, a metadata table (with sample IDs in a column) to auto-subset the data to samples in that table before returning.
If not provided and these_sample_ids is also not provided, the function will return SSM for all samples from the specified seq_type in the bundled metadata.}

\item{this_seq_type}{Default is genome.}

\item{projection}{The projection genome build. Supports hg38 and grch37.}

\item{tool_name}{Optionally specify which tool to report variant from. The default is slms-3, also supports "publication" to return the exact variants as reported in the original papers.}

\item{verbose}{Enable for debugging/noisier output.}

\item{...}{Any additional parameters.}
}
\value{
data frame in MAF format.
}
\description{
Get the SSMs (i.e. load MAF) for a single sample or a collection of samples.
}
\details{
Retrieve a maf for a specific sample or a set of samples.
Either specify the sample IDs of interest with \code{these_sample_ids}.
Or a metadata table subset to the sample IDs of interest with \code{these_samples_metadata}.
}
\examples{
#load a common dependency
library(dplyr)

#Get genome-wide set of mutations from all DLBCL cell lines
cell_line_meta = get_gambl_metadata() \%>\% 
  dplyr::filter(cohort == "DLBCL_cell_lines")

dlbcl_maf = get_ssm_by_samples(these_samples_metadata = cell_line_meta)

}
