% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma_genes_bl_v_latest}
\alias{lymphoma_genes_bl_v_latest}
\title{Lymphoma Genes BL Latest}
\format{
\subsection{\code{lymphoma_genes_bl_v_latest}}{

A data frame withh 128 rows and 11 columns.
\describe{
\item{ensembl_gene_id}{Gene ID in ensembl format.}
\item{Gene}{Gene symbol in Hugo format.}
\item{earliest_support_BL}{Pubmeed ID to associated study.}
\item{curated}{Boolean variable annotating if the event is currated or not.}
\item{aSHM_target_DLBCL}{Boolean variable annotating if the event is an aSHM target in DLBCL or not.}
\item{Currator_comments}{Comments from the currator.}
\item{Enrichment_DLBCL_BL}{Annotates the event if it is enriched in BL or DLBCL.}
\item{frequency_BL_Thomas}{The frequency of which the event was reported in the Thomas study.}
\item{frequency_BL_Panea}{The frequency of which the event was reported in the Panea study.}
\item{n_BL_Panea_original}{Total number of mutated tumors as originally reported in Panea study.}
\item{frequency_BL_Panea_original}{Frequency of mutation as originally reported in Panea study.}
}
}
}
\usage{
lymphoma_genes_bl_v_latest
}
\description{
Genes frequently associated with Burkitt Lymphome (BL). This is the most up-to-date version of this dataset.
}
\keyword{datasets}
