% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma_genes_mcl_v_latest}
\alias{lymphoma_genes_mcl_v_latest}
\title{Lymphoma Genes MCL Latest}
\format{
\subsection{\code{lymphoma_genes_mcl_v_latest}}{

A data frame withh 69 rows and 16 columns.
\describe{
\item{Gene}{Gene ID in Hugo format.}
\item{curated}{Whether this is a manually curated gene.}
\item{Tier}{Tier of gene in this pathology.}
\item{Bea}{Whether this gene was described in Bea study.}
\item{Zhang}{Whether this gene was described in Zhang study.}
\item{Pararajalingam}{Whether this gene was described in Pararajalingam study.}
\item{Nadeu}{Whether this gene was described in Nadeu study.}
\item{Other_support}{Whether there is a support for this gene other than the studies above.}
\item{appx_overall_freq}{Approximate frequency of mutations at this gene in MCL.}
\item{gambl_freq}{Frequency of mutations at this gene in MCL as inferred in GAMBL.}
\item{common_alias}{Alias name for this gene.}
\item{noncoding_driver_suspect}{Whether there are relevant non-coding mutations at this gene.}
\item{aSHM}{Whether this gene is associated with aSHM.}
\item{QC}{QC of a particular gene.}
\item{CIViC}{Whether this gene is reported in the CIViC knowledge base.}
\item{Earliest_support}{The earlist study to describe this gene to be mutated in MCL.}
\item{citekey}{Alphanumeric representation of the citekey where this gene was first described.}
}
}
}
\usage{
lymphoma_genes_mcl_v_latest
}
\description{
Genes frequently associated with Mantle cell lymphoma (MCL).
This is the most up-to-date version of this dataset.
}
\keyword{datasets}
