% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mirage_metrics}
\alias{mirage_metrics}
\title{Mirage Metrics}
\format{
\subsection{\code{mirage_metrics}}{

A data frame with 2376 rows and 17 columns
\describe{
\item{sample_id}{Sample identifier.}
\item{AverageBaseQuality}{Average base quality.}
\item{AverageInsertSize}{Average insert size.}
\item{AverageReadLength}{Average read length.}
\item{PairsOnDiffCHR}{}
\item{TotalReads}{Total reads.}
\item{TotallyUniquelyMapped}{Total uniquely mapped reads.}
\item{TotalUnmappedreads}{Total unmapped reads.}
\item{TotalDuplicatedreads}{Total duplicated reads.}
\item{ProportionReadsDuplicated}{Proportion of duplicated reads.}
\item{ProportionReadsMapped}{Proportion of mapepd reads.}
\item{MeanCorrectedCoverage}{Mean corrected coverage.}
\item{ProportionTargetsNoCoverage}{Proportion of targets with no coverage.}
\item{ProportionCoverage10x}{Proportion of reads with at least 10x coverage.}
\item{study}{Study name}
\item{coding_mutations}{Number of coding mutations.}
\item{ProportionCoverage30x}{Proportions of reads with at least 30x coverage.}
}
}
}
\usage{
mirage_metrics
}
\description{
A data frame with metadata for a collection of GAMBL samples.
}
\keyword{datasets}
